GPU Display: fix compiler warnings

+++ GPU/GPUTracking/display/GPUDisplayMagneticField.cxx
@@ -65,7 +65,7 @@ std::tuple<std::size_t, std::size_t, std::size_t> loadSegments(std::ifstream& fi
   file.read(reinterpret_cast<char*>(&segments.MaxZ), sizeof(segments.MaxZ));
   file.read(reinterpret_cast<char*>(&segments.MultiplicativeFactor), sizeof(segments.MultiplicativeFactor));
 
-  std::int32_t NSegDim1, NSegDim2, NSegDim3;
+  std::size_t NSegDim1, NSegDim2, NSegDim3;
 
   file.read(reinterpret_cast<char*>(&NSegDim1), sizeof(std::int32_t));
   assert(NSegDim1 <= MAX_DIM1_SEGMENTS);
@@ -97,7 +97,7 @@ std::tuple<std::size_t, std::size_t, std::size_t> loadSegments(std::ifstream& fi
 template <std::size_t DIMENSIONS, std::size_t MAX_PARAMETERIZATIONS, std::size_t MAX_ROWS, std::size_t MAX_COLUMNS, std::size_t MAX_COEFFICIENTS>
 std::tuple<std::size_t, std::size_t, std::size_t, std::size_t> loadParams(std::ifstream& file, GPUDisplayMagneticField::ParametrizationUniform<MAX_PARAMETERIZATIONS, MAX_ROWS, MAX_COLUMNS, MAX_COEFFICIENTS>& parametrizations)
 {
-  std::int32_t NParams, NRows, NColumns, NCoefficients;
+  std::size_t NParams, NRows, NColumns, NCoefficients;
   file.read(reinterpret_cast<char*>(&NParams), sizeof(std::int32_t));
   assert(NParams <= (MAX_PARAMETERIZATIONS / DIMENSIONS));
